requirements.txt
absl-py==0.7.1
astor==0.8.0
backports.weakref==1.0.post1
easydict==1.9
enum34==1.1.6
funcsigs==1.0.2
future==0.17.1
futures==3.2.0
gast==0.2.2
grpcio==1.21.1
h5py==2.9.0
Keras-Applications==1.0.6
Keras-Preprocessing==1.0.5
Markdown==3.1.1
mock==3.0.5
numpy==1.16.4
opencv-python==4.1.0.25
Pillow==6.0.0
protobuf==3.8.0
six==1.12.0
tensorboard==1.13.1
tensorflow-estimator==1.13.0
tensorflow-gpu==1.13.1
termcolor==1.1.0
Werkzeug==0.15.4
